fix(src/Plazza/): time is now always a double

+++ include/Interfaces/ProductInterface.hpp
@@ -23,8 +23,8 @@ template <typename Type, typename Size, typename IngredientType> class IProduct 
     [[nodiscard]] virtual Size getSize() const = 0;
     virtual void setSize(Size size) = 0;
 
-    [[nodiscard]] virtual size_t getPreparationTime() const = 0;
-    virtual void setPreparationTime(size_t preparationTime) = 0;
+    [[nodiscard]] virtual double getPreparationTime() const = 0;
+    virtual void setPreparationTime(double preparationTime) = 0;
 
     [[nodiscard]] virtual const std::vector<IngredientType> &getIngredients() const = 0;
     virtual void setIngredients(const std::vector<IngredientType> &ingredients) = 0;
