+++ src/Plazza/Product/Pizza/Pizza.hpp
@@ -17,7 +17,7 @@ namespace Pizzeria
     class Pizza : public AProduct<PizzaType, PizzaSize, PizzaIngredient> {
       public:
         Pizza() = default;
-        Pizza(PizzaType type, PizzaSize size, size_t timeSec, const std::vector<PizzaIngredient> &ingredients);
+        Pizza(PizzaType type, PizzaSize size, double timeSec, const std::vector<PizzaIngredient> &ingredients);
         ~Pizza() override = default;
     };
 } // namespace Pizzeria
