+++ src/Plazza/Restaurant/Restaurant.cpp
@@ -13,7 +13,7 @@ using namespace Pizzeria;
 
 template <typename ProductType, typename ProductSize, typename ProductIngredientType>
 Restaurant<ProductType, ProductSize, ProductIngredientType>::Restaurant(
-    double bakingTime, size_t cooksPerKitchen, size_t restockTime)
+    double bakingTime, size_t cooksPerKitchen, double restockTime)
     : _bakingMultiplier(bakingTime), _cooksPerKitchen(cooksPerKitchen), _restockTime(restockTime), _reception(bakingTime)
 {
 }
